remove simple push/pop pairs

+++ src/bdc/Node.java
@@ -8,12 +8,73 @@ import java.util.Set;
 
 class Node implements InputNode, OutputNode {
 
+    public enum NodeType {
+	INIT,
+
+	POP,
+
+	PUSH,
+
+	STORE_LOCAL,
+
+	LOAD_LOCAL,
+
+	CONSTANT,
+
+	INVOKE,
+
+	LOAD_ELEMENT,
+
+	STORE_ELEMENT,
+
+	CHECKED_CAST,
+
+	INSTANCE_OF,
+
+	MONITOR_ENTER,
+
+	MONITOR_EXIT,
+
+	COMPARE,
+
+	CONVERT,
+
+	BINARY_OPERATION,
+
+	NEGATE,
+
+	SHIFT,
+
+	BITWISE_OPERATION,
+
+	LOAD_STATIC_FIELD,
+
+	STORE_STATIC_FIELD,
+
+	LOAD_FIELD,
+
+	STORE_FIELD,
+
+	NEW_INSTANCE,
+
+	NEW_PRIMITIVE_ARRAY,
+
+	NEW_ARRAY,
+
+	ARRAY_LENGTH,
+
+	TERMINATOR,
+    }
+
+    private final NodeType type;
     private final List<InputPort> input;
     List<Object> description;
     private final List<OutputPort> output;
 
-    public Node(final List<Object> description, final int outputs, final List<? extends OutputPort> input) {
+    public Node(final List<Object> description, final NodeType type, final int outputs,
+	    final List<? extends OutputPort> input) {
 	this.description = description;
+	this.type = type;
 	this.input = new ArrayList<>();
 	this.output = new ArrayList<>();
 	for (int i = 0; i < outputs; i++) {
@@ -26,8 +87,12 @@ class Node implements InputNode, OutputNode {
 	}
     }
 
-    public Node(final List<Object> description, final int outputs, final OutputPort... input) {
-	this(description, outputs, Arrays.asList(input));
+    public Node(final List<Object> description, final NodeType type, final int outputs, final OutputPort... input) {
+	this(description, type, outputs, Arrays.asList(input));
+    }
+
+    public NodeType getType() {
+	return this.type;
     }
 
     @Override
@@ -78,23 +143,26 @@ class Node implements InputNode, OutputNode {
 	}
     }
 
-    public void mergeOutput(final InputNode terminator) {
-	System.out.println(terminator.getInput());
-	System.out.println(getOutput());
+    public static void merge(final InputNode inputNode, final OutputNode outputNode) {
+	final List<? extends InputPort> input = inputNode.getInput();
+	final List<? extends OutputPort> output = outputNode.getOutput();
+	if (input.size() != output.size()) {
+	    throw new IllegalStateException();
+	}
+	for (int i = 0; i < input.size(); i++) {
+	    merge(input.get(i), output.get(i));
+	}
     }
 
-    public static void merge(final InputNode inputNode, final OutputNode outputNode) {
-	if (inputNode instanceof OutputNode) {
-	    Iterables.assertEmpty(((OutputNode) inputNode).getOutput());
-	}
-	if (outputNode instanceof InputNode) {
-	    Iterables.assertEmpty(((InputNode) outputNode).getInput());
-	}
-	final InputPort input = Iterables.getOnlyElement(inputNode.getInput());
-	final OutputPort output = Iterables.getOnlyElement(outputNode.getOutput());
+    public static void merge(final InputPort input, final OutputPort output) {
 	output.replaceWith(input.getSource());
-
 	Iterables.assertEmpty(output.getTargets());
 	input.unlink();
     }
+
+    @Override
+    public String toString() {
+	return getNodeId() + "{" + getType() + "}";
+    }
+
 }
